refactor(content-system): extract element-local wiring into mirrored codec and descriptor classes
