+++ src/Core/Framework/ContentSystem/Layout/Codec/StoredElementCodec.php
@@ -4,18 +4,8 @@ namespace Shopware\Core\Framework\ContentSystem\Layout\Codec;
 
 use Shopware\Core\Framework\ContentSystem\Api\DraftLayoutDecoder;
 use Shopware\Core\Framework\ContentSystem\ContentSystemException;
-use Shopware\Core\Framework\ContentSystem\Hydration\DataContext\ContextType;
 use Shopware\Core\Framework\ContentSystem\Hydration\DataLoader\DataLoaderConfigSerializerProvider;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\ContextConsumer;
 use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\ContextDefinitions;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\ContextProvider;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\BroadcastDistributionConfig;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\DistributionConfig;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\DistributionStrategy;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\IndexedDistributionConfig;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\IteratorDistributionConfig;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\KeyedDistributionConfig;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\SlicedDistributionConfig;
 use Shopware\Core\Framework\ContentSystem\Layout\Element\DataRequirement\DataRequirement;
 use Shopware\Core\Framework\ContentSystem\Layout\Element\StoredElement;
 use Shopware\Core\Framework\ContentSystem\Layout\Element\StoredValue;
@@ -24,7 +14,6 @@ use Shopware\Core\Framework\ContentSystem\Layout\Element\Style\ElementStyle;
 use Shopware\Core\Framework\ContentSystem\Layout\Field\StoredElementListFieldSerializer;
 use Shopware\Core\Framework\ContentSystem\Layout\Scaffolding\VirtualRootWrapper;
 use Shopware\Core\Framework\ContentSystem\Output\Index\ResolvedValueIndexFactory;
-use Shopware\Core\Framework\ContentSystem\Rendering\WiringPlanner;
 use Shopware\Core\Framework\Log\Package;
 
 /**
@@ -38,19 +27,13 @@ use Shopware\Core\Framework\Log\Package;
  * through {@see StoredValue::fromDecoded()} or a typed {@see StoredValue} constructor, and an unknown key is
  * a decode failure rather than a silently dropped field — at the top level, inside a data requirement entry
  * and inside a consumer entry alike, matching the three key sets {@see StoredTreeConstraints} closes on the
- * write path. A provider entry is the one map that does carry extra keys: the declared distribution
- * strategy's own fields ride alongside `type` and `distribution`, and the strategy's config object reads
- * them. Numeric wiring keys are not checked here — the element constructor rejects them, and decode leaves
- * that throw untouched.
+ * write path. Numeric wiring keys are not checked here — the element constructor rejects them, and decode
+ * leaves that throw untouched.
  *
- * Decode also judges an element's context wiring, in two tiers and first hit throws. Per consumer, inside
- * {@see decodeConsumers()}: a `consumerAlias` without `redistribute`, and a `propertyAlias` carrying dot
- * notation. Then, once that map is complete, the element-local tier in
- * {@see rejectInvalidElementWiring()}: landing-key uniqueness across the consumer map, a `redistribute`
- * consumer keyed by a dotted path, and a `redistribute` consumer whose derived provider key an authored
- * provider already holds. The tiers are ordered, not interleaved, so a per-consumer violation anywhere in the
- * element throws before an element-local one. A stored row already carrying an element-local violation is
- * unreadable here rather than repaired.
+ * Both context-wiring maps are decoded by {@see ContextWiringDecoder}, which also owns the consumer entry's
+ * closed key set, the provider entry's deliberately open one, and the two ordered tiers of wiring rules
+ * decode judges. This class delegates all three of its entry points and adds nothing to them; the element-local
+ * tier runs from {@see decodeElement()} once both maps are complete.
  *
  * A malformed structural container is invalid storage and fails decode outright: `dataRequirements`, `slots`,
  * `providesContext`, `acceptsContext`, `style` and `attributedSpecifications` all throw when present as a
@@ -64,12 +47,9 @@ use Shopware\Core\Framework\Log\Package;
  * for malformed input. {@see decodeStyle()} is structurally strict — a malformed option name, value or breakpoint
  * throws like any other malformed container — but stays registry-blind: an option name the registry no longer
  * knows is a well-formed string and still reads verbatim, so removing a style option provider does not make an
- * already-stored layout unreadable; the registry-aware check belongs to the write boundary, not here. A
- * provider entry's already-noted open key set is the other leniency: the declared strategy's own fields ride
- * alongside `type` and `distribution` without decode enforcing the closed set it applies to every other map.
- * Those strategy fields are, in turn, judged by the distribution config's own `fromArray()`, which substitutes
- * its declared default for a field that is absent or null but rejects a present one of the wrong type; decode
- * neither validates nor repeats that judgement itself. Every other malformed value throws.
+ * already-stored layout unreadable; the registry-aware check belongs to the write boundary, not here. The
+ * provider entry's open key set, described on {@see ContextWiringDecoder}, is the other leniency. Every other
+ * malformed value throws.
  *
  * @internal
  */
@@ -108,20 +88,12 @@ final class StoredElementCodec
         'config',
     ];
 
-    /**
-     * @var list<string>
-     */
-    private const CONSUMER_KEYS = [
-        'type',
-        'required',
-        'redistribute',
-        'consumerAlias',
-        'propertyAlias',
-    ];
+    private readonly ContextWiringDecoder $wiring;
 
     public function __construct(
         private readonly DataLoaderConfigSerializerProvider $configProvider,
     ) {
+        $this->wiring = new ContextWiringDecoder();
     }
 
     /**
@@ -175,12 +147,12 @@ final class StoredElementCodec
 
         $properties = $this->decodeProperties($data['properties'] ?? []);
         $slots = $this->decodeSlots($data['slots'] ?? [], $depth);
-        $providers = $this->decodeProviders($data['providesContext'] ?? []);
-        $consumers = $this->decodeConsumers($data['acceptsContext'] ?? []);
+        $providers = $this->wiring->decodeProviders($data['providesContext'] ?? []);
+        $consumers = $this->wiring->decodeConsumers($data['acceptsContext'] ?? []);
         $style = $this->decodeStyle($data['style'] ?? []);
         $attributedSpecifications = $this->decodeAttributedSpecifications($data['attributedSpecifications'] ?? []);
 
-        $this->rejectInvalidElementWiring($consumers, $providers);
+        $this->wiring->rejectInvalidElementWiring($consumers, $providers);
 
         return new StoredElement(
             $id,
@@ -192,51 +164,6 @@ final class StoredElementCodec
             $style,
             $attributedSpecifications,
         );
-    }
-
-    /**
-     * The element-local wiring rules that judge one consumer against another, or the consumer map against the
-     * element's own provider map. They run after {@see decodeConsumers()} has finished the whole map, so a
-     * per-consumer combination violation anywhere in the element throws before any rule here, whatever order
-     * the consumers sit in.
-     *
-     * The three rules are replicated from {@see WiringPlanner}, so a tree the render would reject never
-     * reaches storage; the planner keeps its own copy for writes that bypass the DAL boundary.
-     *
-     * @param array<string, ContextConsumer> $consumers
-     * @param array<string, ContextProvider> $providers
-     */
-    private function rejectInvalidElementWiring(array $consumers, array $providers): void
-    {
-        $holders = [];
-
-        foreach ($consumers as $contextKey => $consumer) {
-            $propertyKey = $consumer->propertyAlias ?? $contextKey;
-
-            $baseKey = str_contains($propertyKey, '.')
-                ? substr($propertyKey, 0, (int) strpos($propertyKey, '.'))
-                : $propertyKey;
-
-            if (\array_key_exists($baseKey, $holders)) {
-                throw ContentSystemException::propertyAliasCollision($baseKey, $holders[$baseKey], $contextKey);
-            }
-
-            $holders[$baseKey] = $contextKey;
-        }
-
-        foreach ($consumers as $contextKey => $consumer) {
-            if (!$consumer->redistribute) {
-                continue;
-            }
-
-            if (str_contains($contextKey, '.')) {
-                throw ContentSystemException::redistributeWithDottedPath($contextKey);
-            }
-
-            if (\array_key_exists($consumer->propertyAlias ?? $contextKey, $providers)) {
-                throw ContentSystemException::redistributeConflict($contextKey);
-            }
-        }
     }
 
     /**
@@ -417,143 +344,6 @@ final class StoredElementCodec
     }
 
     /**
-     * @return array<string, ContextProvider>
-     */
-    private function decodeProviders(mixed $raw): array
-    {
-        if (!\is_array($raw)) {
-            throw ContentSystemException::invalidFieldValueType('providesContext', 'array', get_debug_type($raw));
-        }
-
-        $providers = [];
-        foreach ($raw as $key => $config) {
-            if (!\is_string($key)) {
-                throw ContentSystemException::invalidMapKey('Element context provider map', get_debug_type($key));
-            }
-
-            $path = \sprintf('providesContext[%s]', $key);
-
-            if (!\is_array($config)) {
-                throw ContentSystemException::invalidFieldValueType($path, 'array', get_debug_type($config));
-            }
-
-            $type = $config['type'] ?? null;
-            $contextType = \is_string($type) ? ContextType::tryFrom($type) : null;
-            if ($contextType === null) {
-                throw ContentSystemException::invalidFieldValueType(
-                    $path . '.type',
-                    implode('|', ContextType::values()),
-                    get_debug_type($type)
-                );
-            }
-
-            $distribution = $config['distribution'] ?? null;
-            $strategy = \is_string($distribution) ? DistributionStrategy::tryFrom($distribution) : null;
-            if ($strategy === null) {
-                throw ContentSystemException::invalidFieldValueType(
-                    $path . '.distribution',
-                    implode('|', DistributionStrategy::values()),
-                    get_debug_type($distribution)
-                );
-            }
-
-            $providers[$key] = new ContextProvider(
-                $contextType,
-                $this->distributionConfig($strategy, $this->stringKeyed($config, $path))
-            );
-        }
-
-        return $providers;
-    }
-
-    /**
-     * @param array<string, mixed> $config
-     */
-    private function distributionConfig(DistributionStrategy $strategy, array $config): DistributionConfig
-    {
-        return match ($strategy) {
-            DistributionStrategy::Broadcast => BroadcastDistributionConfig::fromArray($config),
-            DistributionStrategy::Indexed => IndexedDistributionConfig::fromArray($config),
-            DistributionStrategy::Iterator => IteratorDistributionConfig::fromArray($config),
-            DistributionStrategy::Keyed => KeyedDistributionConfig::fromArray($config),
-            DistributionStrategy::Sliced => SlicedDistributionConfig::fromArray($config),
-        };
-    }
-
-    /**
-     * @return array<string, ContextConsumer>
-     */
-    private function decodeConsumers(mixed $raw): array
-    {
-        if (!\is_array($raw)) {
-            throw ContentSystemException::invalidFieldValueType('acceptsContext', 'array', get_debug_type($raw));
-        }
-
-        $consumers = [];
-        foreach ($raw as $key => $config) {
-            if (!\is_string($key)) {
-                throw ContentSystemException::invalidMapKey('Element context consumer map', get_debug_type($key));
-            }
-
-            $path = \sprintf('acceptsContext[%s]', $key);
-
-            if (!\is_array($config)) {
-                throw ContentSystemException::invalidFieldValueType($path, 'array', get_debug_type($config));
-            }
-
-            $this->rejectUnknownKeys($config, self::CONSUMER_KEYS, $path, 'consumer');
-
-            $type = $config['type'] ?? null;
-            $contextType = \is_string($type) ? ContextType::tryFrom($type) : null;
-            if ($contextType === null) {
-                throw ContentSystemException::invalidFieldValueType(
-                    $path . '.type',
-                    implode('|', ContextType::values()),
-                    get_debug_type($type)
-                );
-            }
-
-            $required = $config['required'] ?? null;
-            if (!\is_bool($required)) {
-                throw ContentSystemException::invalidFieldValueType($path . '.required', 'bool', get_debug_type($required));
-            }
-
-            $redistribute = $config['redistribute'] ?? false;
-            if (!\is_bool($redistribute)) {
-                throw ContentSystemException::invalidFieldValueType($path . '.redistribute', 'bool', get_debug_type($redistribute));
-            }
-
-            $consumerAlias = $config['consumerAlias'] ?? null;
-            if ($consumerAlias !== null && !\is_string($consumerAlias)) {
-                throw ContentSystemException::invalidFieldValueType($path . '.consumerAlias', 'string', get_debug_type($consumerAlias));
-            }
-
-            $propertyAlias = $config['propertyAlias'] ?? null;
-            if ($propertyAlias !== null && !\is_string($propertyAlias)) {
-                throw ContentSystemException::invalidFieldValueType($path . '.propertyAlias', 'string', get_debug_type($propertyAlias));
-            }
-
-            if ($consumerAlias !== null && !$redistribute) {
-                throw ContentSystemException::consumerAliasWithoutRedistribute($key);
-            }
-
-            if ($propertyAlias !== null && str_contains($propertyAlias, '.')) {
-                throw ContentSystemException::propertyAliasWithDotNotation($key, $propertyAlias);
-            }
-
-            $consumers[$key] = new ContextConsumer(
-                type: $contextType,
-                required: $required,
-                redistribute: $redistribute,
-                consumerAlias: $consumerAlias,
-                propertyAlias: $propertyAlias,
-            );
-        }
-
-        return $consumers;
-    }
-
-    /**
      * Registry-free structural decode: an option name must be a string, a scalar value is kept verbatim as a
      * flat option, and an array value must be a non-empty map keyed by declared breakpoints with scalar
      * values. Every malformed shape throws rather than being dropped — the write path decodes before the
@@ -657,8 +447,9 @@ final class StoredElementCodec
 
     /**
      * A key set decode closes: a key outside it is a decode failure, so a field the shape does not carry is
-     * never silently dropped on the way to the stored model. The write-path descriptor closes the same three
-     * sets, so neither side admits what the other refuses.
+     * never silently dropped on the way to the stored model. The two sets closed here are the element's own
+     * and the data requirement entry's; {@see ContextWiringDecoder} closes the third, the consumer entry's.
+     * The write-path descriptor closes the same three, so neither side admits what the other refuses.
      *
      * @param array<array-key, mixed> $data
      * @param list<string> $known
@@ -679,8 +470,8 @@ final class StoredElementCodec
     }
 
     /**
-     * A JSON object with a numeric member name decodes to an integer array key, which the config and
-     * distribution factories below cannot take. Reject it here rather than stringifying it back.
+     * A JSON object with a numeric member name decodes to an integer array key, which the data-requirement
+     * config factory cannot take. Reject it here rather than stringifying it back.
      *
      * @param array<array-key, mixed> $map
      *
