+++ src/Core/Framework/ContentSystem/Layout/Codec/StoredTreeConstraints.php
@@ -2,13 +2,6 @@
 
 namespace Shopware\Core\Framework\ContentSystem\Layout\Codec;
 
-use Shopware\Core\Framework\ContentSystem\Hydration\DataContext\ContextType;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\BroadcastDistributionConfig;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\DistributionStrategy;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\IndexedDistributionConfig;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\IteratorDistributionConfig;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\KeyedDistributionConfig;
-use Shopware\Core\Framework\ContentSystem\Layout\Element\Context\Distribution\SlicedDistributionConfig;
 use Shopware\Core\Framework\ContentSystem\Layout\Element\StoredElement;
 use Shopware\Core\Framework\ContentSystem\Layout\Element\Style\Breakpoint;
 use Shopware\Core\Framework\ContentSystem\Layout\Element\Style\Registry\AbstractContentSystemStyleOptionRegistry;
@@ -18,7 +11,6 @@ use Shopware\Core\Framework\Log\Package;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Constraints\All;
 use Symfony\Component\Validator\Constraints\Callback;
-use Symfony\Component\Validator\Constraints\Choice;
 use Symfony\Component\Validator\Constraints\Collection;
 use Symfony\Component\Validator\Constraints\Count;
 use Symfony\Component\Validator\Constraints\NotBlank;
@@ -46,6 +38,10 @@ use Symfony\Component\Validator\Context\ExecutionContextInterface;
  * Tree-global invariants are deliberately absent: this descriptor sees one element at a time and cannot decide
  * whether an id repeats across the forest.
  *
+ * The context-wiring cluster — both wiring maps and the rule spanning them — lives in
+ * {@see StoredTreeWiringConstraints}, which this class composes and whose constraints it attaches at the two
+ * field positions and, for the cross-map rule, whole-element in {@see elementConstraints()}.
+ *
  * Anything this descriptor accepts, {@see StoredElementCodec} must be able to decode, or a payload passes the
  * write and then throws on every later read. Two composition helpers keep that agreement from being forgotten
  * one field at a time: {@see nonNull()} pairs a value's type assertion with its null rejection, and
@@ -57,10 +53,13 @@ use Symfony\Component\Validator\Context\ExecutionContextInterface;
 #[Package('framework')]
 final class StoredTreeConstraints
 {
+    private readonly StoredTreeWiringConstraints $wiring;
+
     public function __construct(
         private readonly AbstractContentSystemStyleOptionRegistry $registry,
         private readonly StyleOptionConstraintDeriver $deriver,
     ) {
+        $this->wiring = new StoredTreeWiringConstraints();
     }
 
     /**
@@ -98,8 +97,8 @@ final class StoredTreeConstraints
                     'properties' => new Optional($this->stringKeyedMap(new Callback($this->validatePropertyValueDepth(...)))),
                     'dataRequirements' => new Optional($this->dataRequirementConstraints()),
                     'slots' => new Optional($this->slotConstraints($depth)),
-                    'providesContext' => new Optional($this->contextProviderConstraints()),
-                    'acceptsContext' => new Optional($this->contextConsumerConstraints()),
+                    'providesContext' => new Optional($this->wiring->contextProviderConstraints()),
+                    'acceptsContext' => new Optional($this->wiring->contextConsumerConstraints()),
                     'style' => new Optional($this->styleConstraints()),
                     'attributedSpecifications' => new Optional($this->stringKeyedMap(...$this->nonNull(new Type('string')))),
                 ],
@@ -111,7 +110,7 @@ final class StoredTreeConstraints
             new PropertyTypeConformance(),
             // Whole-element for the same reason: the rule spans `acceptsContext` and `providesContext`, and
             // neither the `Collection` above nor a per-entry constraint sees both maps.
-            new Callback($this->validateRedistributeProviderConflicts(...)),
+            new Callback($this->wiring->validateRedistributeProviderConflicts(...)),
         ];
     }
 
@@ -333,245 +332,6 @@ final class StoredTreeConstraints
                 ),
             )
         );
-    }
-
-    /**
-     * A provider entry is itself a string-keyed map — the two declared fields plus the declared strategy's
-     * own fields — so it carries the key check its {@see StoredElementCodec} counterpart applies to it.
-     *
-     * @return list<Constraint>
-     */
-    private function contextProviderConstraints(): array
-    {
-        return $this->stringKeyedMap(
-            ...$this->nonNull(
-                new Type('array'),
-                new Callback($this->validateStringKeys(...)),
-                new Collection(
-                    fields: [
-                        'type' => [new NotBlank(), new Choice(choices: ContextType::values())],
-                        'distribution' => [new NotBlank(), new Choice(choices: DistributionStrategy::values())],
-                    ],
-                    allowExtraFields: true,
-                    allowMissingFields: false
-                ),
-                new Callback($this->validateDistributionFields(...)),
-            )
-        );
-    }
-
-    /**
-     * The fields a provider carries beyond `type` and `distribution` depend on the declared strategy, which the
-     * `Collection` above admits as extra fields; each strategy's own constraint set is applied here.
-     */
-    private function validateDistributionFields(mixed $value, ExecutionContextInterface $context): void
-    {
-        if (!\is_array($value) || !isset($value['distribution'])) {
-            return;
-        }
-
-        $configClass = match ($value['distribution']) {
-            'broadcast' => BroadcastDistributionConfig::class,
-            'indexed' => IndexedDistributionConfig::class,
-            'iterator' => IteratorDistributionConfig::class,
-            'keyed' => KeyedDistributionConfig::class,
-            'sliced' => SlicedDistributionConfig::class,
-            default => null,
-        };
-
-        if ($configClass === null) {
-            return;
-        }
-
-        foreach ($configClass::buildConstraints() as $fieldName => $fieldConstraints) {
-            $fieldValue = $value[$fieldName] ?? null;
-
-            foreach ($fieldConstraints as $constraint) {
-                $violations = $context->getValidator()->validate($fieldValue, $constraint);
-
-                foreach ($violations as $violation) {
-                    $context->buildViolation((string) $violation->getMessage())
-                        ->atPath("[$fieldName]")
-                        ->addViolation();
-                }
-            }
-        }
-    }
-
-    /**
-     * @return list<Constraint>
-     */
-    private function contextConsumerConstraints(): array
-    {
-        return [
-            ...$this->stringKeyedMap(
-                ...$this->nonNull(
-                    new Type('array'),
-                    new Collection(
-                        fields: [
-                            'type' => [new NotBlank(), new Choice(choices: ContextType::values())],
-                            'required' => $this->nonNull(new Type('bool')),
-                            'redistribute' => new Optional([new Type('bool')]),
-                            'consumerAlias' => new Optional([new Type('string')]),
-                            'propertyAlias' => new Optional([new Type('string')]),
-                        ],
-                        allowExtraFields: false,
-                        allowMissingFields: false
-                    ),
-                    new Callback($this->validateConsumerAliases(...)),
-                )
-            ),
-            // Map-level, because both rules are judged per entry against the entry's own map key, which a
-            // constraint inside the `All()` above never sees.
-            new Callback($this->validateConsumerLandingKeys(...)),
-            new Callback($this->validateRedistributeKeyShape(...)),
-        ];
-    }
-
-    /**
-     * The two cross-field consumer rules {@see StoredElementCodec} enforces on decode: a consumer alias
-     * renames the context this consumer redistributes, so it means nothing without `redistribute`, and a
-     * property alias names one property on this element, so it carries no dot notation.
-     */
-    private function validateConsumerAliases(mixed $value, ExecutionContextInterface $context): void
-    {
-        if (!\is_array($value)) {
-            return;
-        }
-
-        if (($value['consumerAlias'] ?? null) !== null && ($value['redistribute'] ?? false) !== true) {
-            $context->buildViolation('This value requires "redistribute" to be true.')
-                ->atPath('[consumerAlias]')
-                ->addViolation();
-        }
-
-        $propertyAlias = $value['propertyAlias'] ?? null;
-        if (\is_string($propertyAlias) && str_contains($propertyAlias, '.')) {
-            $context->buildViolation('This value should be a simple property name without dot notation.')
-                ->atPath('[propertyAlias]')
-                ->addViolation();
-        }
-    }
-
-    /**
-     * The landing key a consumer writes its delivered value to is the base segment of
-     * `propertyAlias ?? contextKey`, and two consumers of one element writing the same one would each
-     * overwrite the other. {@see StoredElementCodec} throws on the first collision; here every colliding
-     * consumer is reported, and the first holder of a base key keeps it so a third consumer on that key is
-     * reported too.
-     *
-     * A shape another constraint already reports is skipped rather than reported twice: a non-array map, a
-     * non-array entry, a non-string map key, a non-string `propertyAlias`.
-     */
-    private function validateConsumerLandingKeys(mixed $value, ExecutionContextInterface $context): void
-    {
-        if (!\is_array($value)) {
-            return;
-        }
-
-        $holders = [];
-
-        foreach ($value as $contextKey => $consumer) {
-            if (!\is_string($contextKey) || !\is_array($consumer)) {
-                continue;
-            }
-
-            $propertyAlias = $consumer['propertyAlias'] ?? null;
-            if ($propertyAlias !== null && !\is_string($propertyAlias)) {
-                continue;
-            }
-
-            $propertyKey = $propertyAlias ?? $contextKey;
-            $baseKey = str_contains($propertyKey, '.')
-                ? substr($propertyKey, 0, (int) strpos($propertyKey, '.'))
-                : $propertyKey;
-
-            if (\array_key_exists($baseKey, $holders)) {
-                $context->buildViolation('This consumer writes the property key {{ key }}, which context {{ first }} already writes.')
-                    ->setParameter('{{ key }}', $baseKey)
-                    ->setParameter('{{ first }}', $holders[$baseKey])
-                    ->atPath('[' . $contextKey . ']')
-                    ->addViolation();
-
-                continue;
-            }
-
-            $holders[$baseKey] = $contextKey;
-        }
-    }
-
-    /**
-     * A redistributing consumer becomes a provider keyed by what it writes, and a dotted context key names a
-     * path into a delivered value rather than a key a provider could carry, so the two cannot be combined.
-     * The map key is what carries the dot, which is why this is judged here and not per entry.
-     */
-    private function validateRedistributeKeyShape(mixed $value, ExecutionContextInterface $context): void
-    {
-        if (!\is_array($value)) {
-            return;
-        }
-
-        foreach ($value as $contextKey => $consumer) {
-            if (!\is_string($contextKey) || !\is_array($consumer)) {
-                continue;
-            }
-
-            if (($consumer['redistribute'] ?? false) !== true || !str_contains($contextKey, '.')) {
-                continue;
-            }
-
-            $context->buildViolation('This context key uses dot notation and cannot be redistributed.')
-                ->atPath('[' . $contextKey . '][redistribute]')
-                ->addViolation();
-        }
-    }
-
-    /**
-     * The provider a redistributing consumer derives is keyed by `propertyAlias ?? contextKey`, so an authored
-     * provider already holding that key would be silently overwritten. A `consumerAlias` renames only what
-     * children match on and never enters this comparison.
-     *
-     * The rule spans both wiring maps of one element, which is why it is attached whole-element in
-     * {@see elementConstraints()} rather than beside the consumer map's own rules.
-     */
-    private function validateRedistributeProviderConflicts(mixed $value, ExecutionContextInterface $context): void
-    {
-        if (!\is_array($value)) {
-            return;
-        }
-
-        $consumers = $value['acceptsContext'] ?? [];
-        $providers = $value['providesContext'] ?? [];
-
-        if (!\is_array($consumers) || !\is_array($providers)) {
-            return;
-        }
-
-        foreach ($consumers as $contextKey => $consumer) {
-            if (!\is_string($contextKey) || !\is_array($consumer)) {
-                continue;
-            }
-
-            if (($consumer['redistribute'] ?? false) !== true) {
-                continue;
-            }
-
-            $propertyAlias = $consumer['propertyAlias'] ?? null;
-            if ($propertyAlias !== null && !\is_string($propertyAlias)) {
-                continue;
-            }
-
-            $providerKey = $propertyAlias ?? $contextKey;
-
-            if (!\array_key_exists($providerKey, $providers)) {
-                continue;
-            }
-
-            $context->buildViolation('This consumer redistributes under the provider key {{ key }}, which an authored provider already holds.')
-                ->setParameter('{{ key }}', $providerKey)
-                ->atPath('[acceptsContext][' . $contextKey . '][redistribute]')
-                ->addViolation();
-        }
     }
 
     /**
