+++ src/Core/Framework/ContentSystem/Layout/Element/Context/AGENTS.md
@@ -5,8 +5,8 @@
 ## Constraints
 
 - `redistribute: true` generates virtual provider at runtime — the redistribute derivation in `Rendering/WiringPlanner::plan()`, which runs on the stored tree after the partial prune and before the lowering — never persisted
-- `consumerAlias` requires `redistribute: true` — enforced at `Layout/Codec/StoredElementCodec`, which throws `ContentSystemException::consumerAliasWithoutRedistribute()` on decode, and `Layout/Codec/StoredTreeConstraints`, which reports it as a write-descriptor violation on `[consumerAlias]`
+- `consumerAlias` requires `redistribute: true` — enforced at `Layout/Codec/ContextWiringDecoder` (composed by `StoredElementCodec`), which throws `ContentSystemException::consumerAliasWithoutRedistribute()` on decode, and `Layout/Codec/StoredTreeWiringConstraints` (composed by `StoredTreeConstraints`), which reports it as a write-descriptor violation on `[consumerAlias]`
 - `propertyAlias` does NOT require `redistribute` (independent). It renames only the delivery key: matching and dot-path resolution run on the consumer key, and the alias names where the resolved value lands (`Rendering/ContextDistributor::deliverTo()`)
-- `propertyAlias`: no dots — enforced at the same two sites as `consumerAlias` above; unique per element — the landing key is the base segment of `propertyAlias ?? contextKey`, enforced at `Layout/Codec/StoredElementCodec`, which throws `ContentSystemException::propertyAliasCollision()` on decode, at `Layout/Codec/StoredTreeConstraints`, which reports every colliding consumer as a write-descriptor violation on the consumer's own `acceptsContext` entry (though on the DAL write path the codec's decode throw inside `StoredElementListFieldSerializer::normalize()` precedes this constraint pass, so it fires only where the constraints run without a prior decode), and at the wiring validation in `Rendering/WiringPlanner::plan()`, which runs on the pre-prune forest
-- A `redistribute: true` consumer may not be keyed by a dotted path (`redistributeWithDottedPath`) and its derived provider key `propertyAlias ?? contextKey` may not be one an authored provider already holds (`redistributeConflict`; a `consumerAlias` renames only what children match on and never enters that comparison). Both are enforced at the same three sites as the uniqueness rule above: `StoredElementCodec` throws on decode, `StoredTreeConstraints` reports them on `acceptsContext[<contextKey>][redistribute]` (though, as above, the codec's decode throw precedes this constraint pass on the DAL write path, so it fires only where the constraints run without a prior decode), and `WiringPlanner::plan()` judges the pre-prune forest
+- `propertyAlias`: no dots — enforced at the same two sites as `consumerAlias` above; unique per element — the landing key is the base segment of `propertyAlias ?? contextKey`, enforced at `Layout/Codec/ContextWiringDecoder` (composed by `StoredElementCodec`), which throws `ContentSystemException::propertyAliasCollision()` on decode, at `Layout/Codec/StoredTreeWiringConstraints` (composed by `StoredTreeConstraints`), which reports every colliding consumer as a write-descriptor violation on the consumer's own `acceptsContext` entry (though on the DAL write path the codec's decode throw inside `StoredElementListFieldSerializer::normalize()` precedes this constraint pass, so it fires only where the constraints run without a prior decode), and at the wiring validation in `Rendering/WiringPlanner::plan()`, which runs on the pre-prune forest
+- A `redistribute: true` consumer may not be keyed by a dotted path (`redistributeWithDottedPath`) and its derived provider key `propertyAlias ?? contextKey` may not be one an authored provider already holds (`redistributeConflict`; a `consumerAlias` renames only what children match on and never enters that comparison). Both are enforced at the same three sites as the uniqueness rule above: `ContextWiringDecoder` throws on decode, `StoredTreeWiringConstraints` reports them on `acceptsContext[<contextKey>][redistribute]` (though, as above, the codec's decode throw precedes this constraint pass on the DAL write path, so it fires only where the constraints run without a prior decode), and `WiringPlanner::plan()` judges the pre-prune forest
 - A child-facing delivery key is unique per element, counting both the authored providers (`distributionConfig->getConsumerAlias() ?? providerKey`) and the broadcast providers the redistribute derivation adds from `redistribute` consumers (`consumerAlias ?? contextKey`). `ProviderDeliveryKeyResolver` owns the rule and the derived-key formula, throwing `ContentSystemException::providerDeliveryCollision()`; it indexes the authored providers first, so the earlier of two colliding producers is the one reported as `first`. Both enforcement sites call it — the wiring validation in `Rendering/WiringPlanner::plan()` over the pre-prune forest, and the context walk in `Resolution/AvailableContextResolver` over the target and its ancestors — while which elements each site judges stays that site's own decision
