+++ src/Core/Framework/ContentSystem/Rendering/AGENTS.md
@@ -6,7 +6,7 @@
 
 All `final readonly`; all `@internal` except `RenderedElement`, the public render model.
 
-- `WiringPlanner` - the wiring step preceding the render step; `plan(list<StoredElement> $prePruneForest, list<StoredElement> $prunedTree): list<StoredElement>` — validates the stored forest's context wiring, then derives the redistribute broadcast providers and returns the derived tree. Validation throws; derivation never does. Three element-local rules have a write-side counterpart in `Layout/Codec/{StoredElementCodec,StoredTreeConstraints}`; authoritative for DAL-bypassing trees (migrations, raw SQL)
+- `WiringPlanner` - the wiring step preceding the render step; `plan(list<StoredElement> $prePruneForest, list<StoredElement> $prunedTree): list<StoredElement>` — validates the stored forest's context wiring, then derives the redistribute broadcast providers and returns the derived tree. Validation throws; derivation never does. Three element-local rules have a write-side counterpart in `Layout/Codec/{ContextWiringDecoder,StoredTreeWiringConstraints}`; authoritative for DAL-bypassing trees (migrations, raw SQL)
 - `ElementLowering` - entry point of the render step; `lower(list<StoredElement> $forest, RenderingMode $mode, SalesChannelContext $context, Request $request, RenderingCacheContext $cacheContext): LoweringResult`. Owns the forest-wide data walk and the FULL/SKELETON split
 - `ElementDataResolver` - runs ONE element's data requirements; `resolve(StoredElement $stored, SalesChannelContext $context, Request $request, RenderingCacheContext $cacheContext): array<string, ResolvedLoaderValue>` keyed by requirement key. Mints each value's `Output/Index/LoaderValueIdentity` and applies each loader result's cacheability to the cache context
 - `ContextDeliveryResolver` - owns the forest walk (top-down, required for chained re-providing); `resolve(list<StoredElement> $forest, array<string, array<string, mixed>> $loaderValues): ContextDeliveryIndex`, calling `ContextDistributor` at each parent
